requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai==1.51.0
pydantic==2.9.2
python-dotenv==1.0.1
jinja2==3.1.4
websockets==13.1
jira==3.8.0
httpx==0.27.2
